[CPU info] Split getChipModel over separate .cpp files per platform
