+++ src/src/Helpers/Hardware_device_info.cpp
@@ -373,407 +373,13 @@ uint8_t getChipCores() {
   #endif // ifdef ESP8266
 }
 
-const __FlashStringHelper* getChipModel() {
 #ifdef ESP32
-
-  // https://www.espressif.com/en/products/socs
-  // https://github.com/arendst/Tasmota/blob/1e6b78a957be538cf494f0e2dc49060d1cb0fe8b/tasmota/support_esp.ino#L579
-
-  /*
-     Source: esp_chip_info.h & esptool.py  ('IMAGE_CHIP_ID') https://github.com/espressif/esptool/search?q=IMAGE_CHIP_ID
-
-      typedef enum {
-          CHIP_ESP32      = 1,  //!< ESP32
-          CHIP_ESP32S2    = 2,  //!< ESP32-S2
-          CHIP_ESP32S3_b  = 4,  //!< ESP32-S3(beta2)
-          CHIP_ESP32S3    = 9,  //!< ESP32-S3
-          CHIP_ESP32C3    = 5,  //!< ESP32-C3
-          CHIP_ESP32C2    = 12, //!< ESP32-C2
-          CHIP_ESP32C5    = 17, //!< ESP32-C5 beta3 (MPW)
-          CHIP_ESP32C5    = 23, //!< ESP32-C5 MP
-          CHIP_ESP32C6_b  = 7,  //!< ESP32-C6(beta)
-          CHIP_ESP32C6    = 13, //!< ESP32-C6
-          CHIP_ESP32C61   = 20, //!< ESP32-C61
-          CHIP_ESP32H2_b1 = 10, //!< ESP32-H2(beta1)
-          CHIP_ESP32H2_b2 = 14, //!< ESP32-H2(beta2)
-          CHIP_ESP32H2    = 16, //!< ESP32-H2
-          CHIP_ESP32P4    = 18, //!< ESP32-P4
-      } esp_chip_model_t;
-
-      // Chip feature flags, used in esp_chip_info_t
-   #define CHIP_FEATURE_EMB_FLASH      BIT(0)      //!< Chip has embedded flash memory
-   #define CHIP_FEATURE_WIFI_BGN       BIT(1)      //!< Chip has 2.4GHz WiFi
-   #define CHIP_FEATURE_BLE            BIT(4)      //!< Chip has Bluetooth LE
-   #define CHIP_FEATURE_BT             BIT(5)      //!< Chip has Bluetooth Classic
-   #define CHIP_FEATURE_IEEE802154     BIT(6)      //!< Chip has IEEE 802.15.4
-   #define CHIP_FEATURE_EMB_PSRAM      BIT(7)      //!< Chip has embedded psram
-
-      // The structure represents information about the chip
-      typedef struct {
-          esp_chip_model_t model;  //!< chip model, one of esp_chip_model_t
-          uint32_t features;       //!< bit mask of CHIP_FEATURE_x feature flags
-          uint8_t cores;           //!< number of CPU cores
-          uint8_t revision;        //!< chip revision number
-      } esp_chip_info_t;
-   */
-
-  esp_chip_info_t chip_info;
-  esp_chip_info(&chip_info);
-
-  uint32_t chip_model    = chip_info.model;
-  const uint32_t chip_revision =
-#if ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(5, 0, 0)
-    chip_info.revision / 100;
-#else
-    chip_info.revision;
+bool isESP8285() {
+  return false;
+}
 #endif
 
-  uint32_t pkg_version = 0;
-# if (ESP_IDF_VERSION_MAJOR >= 5)
-  pkg_version = bootloader_common_get_chip_ver_pkg();
-# endif // if (ESP_IDF_VERSION_MAJOR >= 5)
 
-  //  uint32_t chip_revision = ESP.getChipRevision();
-  bool rev3 = (3 == chip_revision);
-
-  //  bool single_core = (1 == ESP.getChipCores());
-  bool single_core = (1 == chip_info.cores);
-
-  if (chip_model < 2) { // ESP32
-# if CONFIG_IDF_TARGET_ESP32
-
-    /* esptool:
-        def get_pkg_version(self):
-            word3 = self.read_efuse(3)
-            pkg_version = (word3 >> 9) & 0x07
-            pkg_version += ((word3 >> 2) & 0x1) << 3
-            return pkg_version
-     */
-#  if ESP_IDF_VERSION_MAJOR < 5
-    uint32_t chip_ver = REG_GET_FIELD(EFUSE_BLK0_RDATA3_REG, EFUSE_RD_CHIP_VER_PKG);
-    pkg_version = chip_ver & 0x7;
-#  else // if ESP_IDF_VERSION_MAJOR < 5
-    uint32_t chip_ver = REG_GET_FIELD(EFUSE_BLK0_RDATA3_REG, EFUSE_RD_CHIP_PACKAGE);
-#  endif // if ESP_IDF_VERSION_MAJOR < 5
-    
-
-    //    AddLog(LOG_LEVEL_DEBUG_MORE, PSTR("HDW: ESP32 Model %d, Revision %d, Core %d, Package %d"), chip_info.model, chip_revision,
-    // chip_info.cores, chip_ver);
-
-    switch (pkg_version) {
-      case EFUSE_RD_CHIP_VER_PKG_ESP32D0WDQ6:
-
-        if (single_core) { return F("ESP32-S0WDQ6"); }    // Max 240MHz, Single core, QFN 6*6
-        else if (rev3)   { return F("ESP32-D0WDQ6-V3"); } // Max 240MHz, Dual core, QFN 6*6
-        else {             return F("ESP32-D0WDQ6"); }    // Max 240MHz, Dual core, QFN 6*6
-      case EFUSE_RD_CHIP_VER_PKG_ESP32D0WDQ5:
-
-        if (single_core) { return F("ESP32-S0WD"); }   // Max 160MHz, Single core, QFN 5*5, ESP32-SOLO-1, ESP32-DevKitC
-        else if (rev3)   { return F("ESP32-D0WDQ5-V3"); } // Max 240MHz, Dual core, QFN 5*5, ESP32-WROOM-32E, ESP32_WROVER-E, ESP32-DevKitC
-        else {             return F("ESP32-D0WDQ5"); }    // Max 240MHz, Dual core, QFN 5*5, ESP32-WROOM-32D, ESP32_WROVER-B, ESP32-DevKitC
-      case EFUSE_RD_CHIP_VER_PKG_ESP32D2WDQ5:
-        return F("ESP32-D2WDQ5");                      // Max 160MHz, Dual core, QFN 5*5, 2MB embedded flash
-      case 3:
-
-        if (single_core) { return F("ESP32-S0WD-OEM"); } // Max 160MHz, Single core, QFN 5*5, Xiaomi Yeelight
-        else {             return F("ESP32-D0WD-OEM"); } // Max 240MHz, Dual core, QFN 5*5
-      case EFUSE_RD_CHIP_VER_PKG_ESP32U4WDH:
-        return F("ESP32-U4WDH");                         // Max 160MHz, Single core, QFN 5*5, 4MB embedded flash, ESP32-MINI-1,
-      // ESP32-DevKitM-1
-      case EFUSE_RD_CHIP_VER_PKG_ESP32PICOD4:
-
-        if (rev3)        { return F("ESP32-PICO-V3"); } // Max 240MHz, Dual core, LGA 7*7, ESP32-PICO-V3-ZERO, ESP32-PICO-V3-ZERO-DevKit
-        else {             return F("ESP32-PICO-D4"); } // Max 240MHz, Dual core, LGA 7*7, 4MB embedded flash, ESP32-PICO-KIT
-      case EFUSE_RD_CHIP_VER_PKG_ESP32PICOV302:
-        return F("ESP32-PICO-V3-02");                    // Max 240MHz, Dual core, LGA 7*7, 8MB embedded flash, 2MB embedded PSRAM,
-                                                         // ESP32-PICO-MINI-02, ESP32-PICO-DevKitM-2
-      case EFUSE_RD_CHIP_VER_PKG_ESP32D0WDR2V3:
-        return F("ESP32-D0WDR2-V3");
-    }
-# endif // if CONFIG_IDF_TARGET_ESP32
-    return F("ESP32");
-  }
-  else if (CHIP_ESP32S2 == chip_model) { // ESP32-S2
-# ifdef CONFIG_IDF_TARGET_ESP32S2
-
-    /* esptool:
-        def get_flash_version(self):
-            num_word = 3
-            block1_addr = self.EFUSE_BASE + 0x044
-            word3 = self.read_reg(block1_addr + (4 * num_word))
-            pkg_version = (word3 >> 21) & 0x0F
-            return pkg_version
-        def get_psram_version(self):
-            num_word = 3
-            block1_addr = self.EFUSE_BASE + 0x044
-            word3 = self.read_reg(block1_addr + (4 * num_word))
-                pkg_version = (word3 >> 28) & 0x0F
-                return pkg_version
-     */
-    uint32_t chip_ver  = REG_GET_FIELD(EFUSE_RD_MAC_SPI_SYS_3_REG, EFUSE_FLASH_VERSION);
-    uint32_t psram_ver = REG_GET_FIELD(EFUSE_RD_MAC_SPI_SYS_3_REG, EFUSE_PSRAM_VERSION);
-    pkg_version = (chip_ver & 0xF) + ((psram_ver & 0xF) * 100);
-
-    //    AddLog(LOG_LEVEL_DEBUG_MORE, PSTR("HDW: ESP32 Model %d, Revision %d, Core %d, Package %d"), chip_info.model, chip_revision,
-    // chip_info.cores, chip_ver);
-
-    switch (pkg_version) {
-      case 0:              return F("ESP32-S2");      // Max 240MHz, Single core, QFN 7*7, ESP32-S2-WROOM, ESP32-S2-WROVER,
-      // ESP32-S2-Saola-1, ESP32-S2-Kaluga-1
-      case 1:              return F("ESP32-S2FH2");   // Max 240MHz, Single core, QFN 7*7, 2MB embedded flash, ESP32-S2-MINI-1,
-      // ESP32-S2-DevKitM-1
-      case 2:              return F("ESP32-S2FH4");   // Max 240MHz, Single core, QFN 7*7, 4MB embedded flash
-      case 3:              return F("ESP32-S2FN4R2"); // Max 240MHz, Single core, QFN 7*7, 4MB embedded flash, 2MB embedded PSRAM, ,
-      // ESP32-S2-MINI-1U, ESP32-S2-DevKitM-1U
-      case 100:            return F("ESP32-S2R2");
-      case 102:            return F("ESP32-S2FNR2");  // Max 240MHz, Single core, QFN 7*7, 4MB embedded flash, 2MB embedded PSRAM, , Lolin
-                                                      // S2 mini
-    }
-
-# endif // CONFIG_IDF_TARGET_ESP32S2
-    return F("ESP32-S2");
-  }
-  else if (CHIP_ESP32C3 == chip_model) { // ESP32-C3
-# ifdef CONFIG_IDF_TARGET_ESP32C3
-
-    /* esptool:
-        def get_pkg_version(self):
-            num_word = 3
-            block1_addr = self.EFUSE_BASE + 0x044
-            word3 = self.read_reg(block1_addr + (4 * num_word))
-            pkg_version = (word3 >> 21) & 0x0F
-            return pkg_version
-     */
-    uint32_t chip_ver = REG_GET_FIELD(EFUSE_RD_MAC_SPI_SYS_3_REG, EFUSE_PKG_VERSION);
-    pkg_version = chip_ver & 0x7;
-
-    //    pkg_version = esp_efuse_get_pkg_ver();
-
-    //    AddLog(LOG_LEVEL_DEBUG_MORE, PSTR("HDW: ESP32 Model %d, Revision %d, Core %d, Package %d"), chip_info.model, chip_revision,
-    // chip_info.cores, chip_ver);
-
-    switch (pkg_version) {
-      case 0:              return F("ESP32-C3");    // Max 160MHz, Single core, QFN 5*5, ESP32-C3-WROOM-02, ESP32-C3-DevKitC-02
-      //        case 1:              return F("ESP32-C3FH4");        // Max 160MHz, Single core, QFN 5*5, 4MB embedded flash,
-      // ESP32-C3-MINI-1, ESP32-C3-DevKitM-1
-      case 1:              return F("ESP8685");     // Max 160MHz, Single core, QFN 5*5, 4MB embedded flash, ESP32-C3-MINI-1,
-      // ESP32-C3-DevKitM-1
-      case 2:              return F("ESP32-C3 AZ"); // QFN32
-      case 3:              return F("ESP8686");     // QFN24
-    }
-# endif // CONFIG_IDF_TARGET_ESP32C3
-    return F("ESP32-C3");
-  }
-  else if ((CHIP_ESP32S3 == chip_model) || // ESP32-S3
-           (4 == chip_model) ||            // ESP32-S3(beta2)
-           (6 == chip_model)) {            // ESP32-S3(beta3)
-    /*
-       ESP32-S3 Series
-       - 32-bit MCU & 2.4 GHz Wi-Fi & Bluetooth 5 (LE)
-       - Xtensa® 32-bit LX7 dual-core processor that operates at up to 240 MHz
-       - 512 KB of SRAM and 384 KB of ROM on the chip, and SPI, Dual SPI, Quad SPI, Octal SPI, QPI, and OPI interfaces that allow connection
-          to flash and external RAM
-       - Additional support for vector instructions in the MCU, which provides acceleration for neural network computing and signal
-          processing workloads
-       - Peripherals include 45 programmable GPIOs, SPI, I2S, I2C, PWM, RMT, ADC and UART, SD/MMC host and TWAITM
-       - Reliable security features ensured by RSA-based secure boot, AES-XTS-based flash encryption, the innovative digital signature and
-          the HMAC peripheral, “World Controller”
-     */
-
-    /*
-
-            efuse_reg.h:
-            EFUSE_RD_MAC_SPI_SYS_0_REG = block1_addr
-            EFUSE_RD_MAC_SPI_SYS_3_REG = block1_addr + (4 * num_word)) // (num_word = 3)
-
-     */
-
-# ifdef CONFIG_IDF_TARGET_ESP32S3
-#  if (ESP_IDF_VERSION_MAJOR >= 5)
-    pkg_version = REG_GET_FIELD(EFUSE_RD_MAC_SPI_SYS_3_REG, EFUSE_PKG_VERSION);
-
-    switch (pkg_version) {
-      case 0:              return F("ESP32-S3");        // QFN56
-      case 1:              return F("ESP32-S3-PICO-1"); // LGA56
-    }
-#  endif // if (ESP_IDF_VERSION_MAJOR >= 5)
-# endif // CONFIG_IDF_TARGET_ESP32S3
-
-    return F("ESP32-S3");      // Max 240MHz, Dual core, QFN 7*7, ESP32-S3-WROOM-1, ESP32-S3-DevKitC-1
-  }
-  else if (12 == chip_model) { // ESP32-C2 = ESP8684 if embedded flash
-    /*
-       ESP32-C2 Series
-       - 32-bit RISC-V MCU & 2.4 GHz Wi-Fi & Bluetooth 5 (LE)
-       - 32-bit RISC-V single-core processor that operates at up to 120 MHz
-       - State-of-the-art power and RF performance
-       - 576 KB ROM, 272 KB SRAM (16 KB for cache) on the chip
-       - 14 programmable GPIOs: SPI, UART, I2C, LED PWM controller, General DMA controller (GDMA), SAR ADC, Temperature sensor
-     */
-# ifdef CONFIG_IDF_TARGET_ESP32C2
-
-    switch (pkg_version) {
-      case 0:              return F("ESP32-C2");
-      case 1:              return F("ESP32-C2");
-    }
-# endif // CONFIG_IDF_TARGET_ESP32C2
-    return F("ESP32-C2");
-  }
-  else if ((7 == chip_model) ||  // ESP32-C6(beta)
-           (13 == chip_model)) { // ESP32-C6
-# ifdef CONFIG_IDF_TARGET_ESP32C6
-
-    /* esptool:
-        def get_pkg_version(self):
-            num_word = 3
-            block1_addr = self.EFUSE_BASE + 0x044
-            word3 = self.read_reg(block1_addr + (4 * num_word))
-            pkg_version = (word3 >> 21) & 0x0F
-            return pkg_version
-     */
-    uint32_t chip_ver = REG_GET_FIELD(EFUSE_RD_MAC_SPI_SYS_3_REG, EFUSE_PKG_VERSION);
-    pkg_version = chip_ver & 0x7;
-
-    //    pkg_version = esp_efuse_get_pkg_ver();
-
-    //    AddLog(LOG_LEVEL_DEBUG_MORE, PSTR("HDW: ESP32 Model %d, Revision %d, Core %d, Package %d"), chip_info.model, chip_revision,
-    // chip_info.cores, chip_ver);
-
-    switch (pkg_version) {
-      case 0: return F("ESP32-C6");
-      case 1: return F("ESP32-C6FH4 (QFN32)");
-    }
-# endif // CONFIG_IDF_TARGET_ESP32C6
-    return F("ESP32-C6");
-  }
-  else if ((10 == chip_model) ||            // ESP32-H2(beta1)
-           (14 == chip_model) ||           // ESP32-H2(beta2)
-           (CHIP_ESP32H2 == chip_model)) { // ESP32-H2
-# ifdef CONFIG_IDF_TARGET_ESP32H2
-
-    /* esptool:
-        def get_pkg_version(self):
-            num_word = 3
-            block1_addr = self.EFUSE_BASE + 0x044
-            word3 = self.read_reg(block1_addr + (4 * num_word))
-            pkg_version = (word3 >> 21) & 0x0F
-            return pkg_version
-     */
-    uint32_t chip_ver = REG_GET_FIELD(EFUSE_RD_MAC_SPI_SYS_3_REG, EFUSE_PKG_VERSION);
-    pkg_version = chip_ver & 0x7;
-
-    //    pkg_version = esp_efuse_get_pkg_ver();
-
-    //    AddLog(LOG_LEVEL_DEBUG_MORE, PSTR("HDW: ESP32 Model %d, Revision %d, Core %d, Package %d"), chip_info.model, chip_revision,
-    // chip_info.cores, chip_ver);
-
-    switch (pkg_version) {
-      case 0:              return F("ESP32-H2");
-    }
-# endif // CONFIG_IDF_TARGET_ESP32H2
-    return F("ESP32-H2");
-  }
-  else if (18 == chip_model) {  // ESP32-P4
-# ifdef CONFIG_IDF_TARGET_ESP32P4
-
-    switch (pkg_version) {
-      case 0:              return F("ESP32-P4");
-    }
-# endif // CONFIG_IDF_TARGET_ESP32P4
-    return F("ESP32-P4");
-  }
-  else if (17 == chip_model) {  // ESP32-C5 beta3 (MPW)
-    return F("ESP32-C5 beta3");
-  }
-  else if (23 == chip_model) {  // ESP32-C5 MP
-    return F("ESP32-C5");
-  }
-  else if (20 == chip_model) {  // ESP32-C61
-    return F("ESP32-C61");
-  }
-
-  return F("ESP32");
-#elif defined(ESP8266)
-  uint32_t pkg_version{};
-  bool     high_temp_version{};
-
-  if (isESP8285(pkg_version, high_temp_version)) {
-    switch (pkg_version) {
-      case 1:
-        return (high_temp_version)
-          ? F("ESP8285H08") // 1M flash
-          : F("ESP8285N08");
-      case 2:
-        return (high_temp_version)
-          ? F("ESP8285H16") // 2M flash
-          : F("ESP8285N16");
-      case 4:
-        return (high_temp_version)
-          ? F("ESP8285H32") // 4M flash
-          : F("ESP8285N32");
-    }
-    return F("ESP8285");
-  }
-  return F("ESP8266EX");
-#endif // ifdef ESP32
-  return F("Unknown");
-}
-
-bool isESP8285(uint32_t& pkg_version, bool& high_temp_version)
-{
-  // Original code from Tasmota:
-  // https://github.com/arendst/Tasmota/blob/62675a37a0e7b46283e2fdfe459bb8fd29d1cc2a/tasmota/tasmota_support/support_esp.ino#L151
-
-  /*
-     ESP8266 SoCs
-     - 32-bit MCU & 2.4 GHz Wi-Fi
-     - High-performance 160 MHz single-core CPU
-     - +19.5 dBm output power ensures a good physical range
-     - Sleep current is less than 20 μA, making it suitable for battery-powered and wearable-electronics applications
-     - Peripherals include UART, GPIO, I2C, I2S, SDIO, PWM, ADC and SPI
-   */
-
-  // esptool.py get_efuses
-  uint32_t efuse0 = *(uint32_t *)(0x3FF00050);
-
-  //  uint32_t efuse1 = *(uint32_t*)(0x3FF00054);
-  uint32_t efuse2 = *(uint32_t *)(0x3FF00058);
-  uint32_t efuse3 = *(uint32_t *)(0x3FF0005C);
-
-  bool r0_4  = efuse0 & (1 << 4);    // ESP8285
-  bool r2_16 = efuse2 & (1 << 16);   // ESP8285
-
-  if (r0_4 || r2_16) {               // ESP8285
-    //                                                              1M 2M 2M 4M flash size
-    //   r0_4                                                       1  1  0  0
-    bool r3_25 = efuse3 & (1 << 25); // flash matrix  0  0  1  1
-    bool r3_26 = efuse3 & (1 << 26); // flash matrix  0  1  0  1
-    bool r3_27 = efuse3 & (1 << 27); // flash matrix  0  0  0  0
-    pkg_version = 0;
-
-    if (!r3_27) {
-      if (r0_4 && !r3_25) {
-        pkg_version = (r3_26) ? 2 : 1;
-      }
-      else if (!r0_4 && r3_25) {
-        pkg_version = (r3_26) ? 4 : 2;
-      }
-    }
-    high_temp_version = efuse0 & (1 << 5); // Max flash temperature (0 = 85C, 1 = 105C)
-    return true;
-  }
-  return false;
-}
-
-bool isESP8285() {
-  #ifdef ESP8266
-  uint32_t pkg_version{};
-  bool     high_temp_version{};
-  return isESP8285(pkg_version, high_temp_version);
-  #else // ifdef ESP8266
-  return false;
-  #endif // ifdef ESP8266
-}
 
 String getChipRevision() {
   static uint16_t rev = 0;
@@ -878,5 +484,75 @@ bool CanUsePSRAM() {
 # endif // if CONFIG_IDF_TARGET_ESP32
   return true;
 }
+
+# ifdef ESP32
+const __FlashStringHelper* getChipModel()
+{
+  
+  // https://www.espressif.com/en/products/socs
+  // https://github.com/arendst/Tasmota/blob/1e6b78a957be538cf494f0e2dc49060d1cb0fe8b/tasmota/support_esp.ino#L579
+
+  /*
+     Source: esp_chip_info.h & esptool.py  ('IMAGE_CHIP_ID') https://github.com/espressif/esptool/search?q=IMAGE_CHIP_ID
+
+     typedef enum {
+     CHIP_ESP32      = 1,  //!< ESP32
+     CHIP_ESP32S2    = 2,  //!< ESP32-S2
+     CHIP_ESP32S3_b  = 4,  //!< ESP32-S3(beta2)
+     CHIP_ESP32S3    = 9,  //!< ESP32-S3
+     CHIP_ESP32C3    = 5,  //!< ESP32-C3
+     CHIP_ESP32C2    = 12, //!< ESP32-C2
+     CHIP_ESP32C5    = 17, //!< ESP32-C5 beta3 (MPW)
+     CHIP_ESP32C5    = 23, //!< ESP32-C5 MP
+     CHIP_ESP32C6_b  = 7,  //!< ESP32-C6(beta)
+     CHIP_ESP32C6    = 13, //!< ESP32-C6
+     CHIP_ESP32C61   = 20, //!< ESP32-C61
+     CHIP_ESP32H2_b1 = 10, //!< ESP32-H2(beta1)
+     CHIP_ESP32H2_b2 = 14, //!< ESP32-H2(beta2)
+     CHIP_ESP32H2    = 16, //!< ESP32-H2
+     CHIP_ESP32P4    = 18, //!< ESP32-P4
+     } esp_chip_model_t;
+
+     // Chip feature flags, used in esp_chip_info_t
+   #define CHIP_FEATURE_EMB_FLASH      BIT(0)      //!< Chip has embedded flash memory
+   #define CHIP_FEATURE_WIFI_BGN       BIT(1)      //!< Chip has 2.4GHz WiFi
+   #define CHIP_FEATURE_BLE            BIT(4)      //!< Chip has Bluetooth LE
+   #define CHIP_FEATURE_BT             BIT(5)      //!< Chip has Bluetooth Classic
+   #define CHIP_FEATURE_IEEE802154     BIT(6)      //!< Chip has IEEE 802.15.4
+   #define CHIP_FEATURE_EMB_PSRAM      BIT(7)      //!< Chip has embedded psram
+
+     // The structure represents information about the chip
+     typedef struct {
+     esp_chip_model_t model;  //!< chip model, one of esp_chip_model_t
+     uint32_t features;       //!< bit mask of CHIP_FEATURE_x feature flags
+     uint8_t cores;           //!< number of CPU cores
+     uint8_t revision;        //!< chip revision number
+     } esp_chip_info_t;
+   */
+
+  esp_chip_info_t chip_info;
+  esp_chip_info(&chip_info);
+
+  uint32_t chip_model          = chip_info.model;
+  const uint32_t chip_revision =
+#  if ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(5, 0, 0)
+    chip_info.revision / 100;
+#  else // if ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(5, 0, 0)
+    chip_info.revision;
+#  endif // if ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(5, 0, 0)
+
+  uint32_t pkg_version = 0;
+#  if (ESP_IDF_VERSION_MAJOR >= 5)
+  pkg_version = bootloader_common_get_chip_ver_pkg();
+#  endif // if (ESP_IDF_VERSION_MAJOR >= 5)
+
+  //  bool single_core = (1 == ESP.getChipCores());
+  bool single_core = (1 == chip_info.cores);
+
+
+  return getChipModel(chip_model, chip_revision, pkg_version, single_core);
+}
+#endif
+
 
 #endif  // ESP32
