+++ src/src/Helpers/Hardware_device_info_ESP32C6.cpp
@@ -6,6 +6,9 @@
 # include <soc/efuse_reg.h>
 # include <soc/spi_reg.h>
 # include <soc/rtc.h>
+# include <esp_chip_info.h>
+# include <bootloader_common.h>
+
 
 bool isFlashInterfacePin_ESPEasy(int gpio) {
   // FIXME TD-er: Must know whether we have internal or external flash
@@ -59,5 +62,37 @@ bool isPSRAMInterfacePin(int gpio) {
 }
 
 # endif // ifndef isPSRAMInterfacePin
+
+
+const __FlashStringHelper* getChipModel(uint32_t chip_model, uint32_t chip_revision, uint32_t pkg_version, bool single_core)
+{
+  if ((7 == chip_model) ||  // ESP32-C6(beta)
+      (13 == chip_model)) { // ESP32-C6
+    /* esptool:
+        def get_pkg_version(self):
+            num_word = 3
+            block1_addr = self.EFUSE_BASE + 0x044
+            word3 = self.read_reg(block1_addr + (4 * num_word))
+            pkg_version = (word3 >> 21) & 0x0F
+            return pkg_version
+     */
+    uint32_t chip_ver = REG_GET_FIELD(EFUSE_RD_MAC_SPI_SYS_3_REG, EFUSE_PKG_VERSION);
+    pkg_version = chip_ver & 0x7;
+
+    //    pkg_version = esp_efuse_get_pkg_ver();
+
+    //    AddLog(LOG_LEVEL_DEBUG_MORE, PSTR("HDW: ESP32 Model %d, Revision %d, Core %d, Package %d"), chip_info.model, chip_revision,
+    // chip_info.cores, chip_ver);
+
+    switch (pkg_version) {
+      case 0: return F("ESP32-C6");
+      case 1: return F("ESP32-C6FH4 (QFN32)");
+    }
+
+    return F("ESP32-C6");
+  }
+
+  return F("Unknown");
+}
 
 #endif // ifdef ESP32C6
