+++ src/src/Helpers/Hardware_device_info_ESP32S3.cpp
@@ -5,10 +5,12 @@
 // See:
 // - https://github.com/espressif/esptool/blob/master/esptool/targets/esp32s3.py
 // - https://github.com/tasmota/esp-idf/blob/206ce4b7f875bf5568ba47aba23f4b28e81b0574/components/efuse/esp32s3/esp_efuse_table.csv#L203-L208
-  # include <soc/soc.h>
-  # include <soc/efuse_reg.h>
-  # include <soc/spi_reg.h>
-  # include <soc/rtc.h>
+# include <soc/soc.h>
+# include <soc/efuse_reg.h>
+# include <soc/spi_reg.h>
+# include <soc/rtc.h>
+# include <esp_chip_info.h>
+# include <bootloader_common.h>
 
 // Flash data lines:
 // https://github.com/tasmota/esp-idf/blob/206ce4b7f875bf5568ba47aba23f4b28e81b0574/components/efuse/esp32s3/esp_efuse_table.csv#L175
@@ -16,30 +18,30 @@
 /** EFUSE_FLASH_CAP : R; bitpos: [29:27]; default: 0;
  * register: EFUSE_RD_MAC_SPI_SYS_3_REG
  */
-  # define EFUSE_FLASH_CAP    0x00000007U
-  # define EFUSE_FLASH_CAP_M  (EFUSE_FLASH_CAP_V << EFUSE_FLASH_CAP_S)
-  # define EFUSE_FLASH_CAP_V  0x00000007U
-  # define EFUSE_FLASH_CAP_S  27
+# define EFUSE_FLASH_CAP    0x00000007U
+# define EFUSE_FLASH_CAP_M  (EFUSE_FLASH_CAP_V << EFUSE_FLASH_CAP_S)
+# define EFUSE_FLASH_CAP_V  0x00000007U
+# define EFUSE_FLASH_CAP_S  27
 
 
 /** EFUSE_FLASH_VENDOR : R; bitpos: [2:0]; default: 0;
  * register: EFUSE_RD_MAC_SPI_SYS_4_REG
  */
-  # define EFUSE_FLASH_VENDOR    0x00000007U
-  # define EFUSE_FLASH_VENDOR_M  (EFUSE_FLASH_VENDOR_V << EFUSE_FLASH_VENDOR_S)
-  # define EFUSE_FLASH_VENDOR_V  0x00000007U
-  # define EFUSE_FLASH_VENDOR_S  0
+# define EFUSE_FLASH_VENDOR    0x00000007U
+# define EFUSE_FLASH_VENDOR_M  (EFUSE_FLASH_VENDOR_V << EFUSE_FLASH_VENDOR_S)
+# define EFUSE_FLASH_VENDOR_V  0x00000007U
+# define EFUSE_FLASH_VENDOR_S  0
 
 
 /*
-    switch (flash_vendor)
-    {
-      case 1: features += F("(XMC)"); break;
-      case 2: features += F("(GD)"); break;
-      case 3: features += F("(FM)"); break;
-      case 4: features += F("(TT)"); break;
-      case 5: features += F("(BY)"); break;
-    }
+   switch (flash_vendor)
+   {
+   case 1: features += F("(XMC)"); break;
+   case 2: features += F("(GD)"); break;
+   case 3: features += F("(FM)"); break;
+   case 4: features += F("(TT)"); break;
+   case 5: features += F("(BY)"); break;
+   }
  */
 bool isFlashInterfacePin_ESPEasy(int gpio) {
   // GPIO-26 ... 32: SPI flash and PSRAM
@@ -64,26 +66,26 @@ int32_t getEmbeddedFlashSize()
 /** EFUSE_PSRAM_CAP : R; bitpos: [4:3]; default: 0;
  * register: EFUSE_RD_MAC_SPI_SYS_4_REG
  */
-  # define EFUSE_PSRAM_CAP    0x00000003U
-  # define EFUSE_PSRAM_CAP_M  (EFUSE_PSRAM_CAP_V << EFUSE_PSRAM_CAP_S)
-  # define EFUSE_PSRAM_CAP_V  0x00000003U
-  # define EFUSE_PSRAM_CAP_S  3
+# define EFUSE_PSRAM_CAP    0x00000003U
+# define EFUSE_PSRAM_CAP_M  (EFUSE_PSRAM_CAP_V << EFUSE_PSRAM_CAP_S)
+# define EFUSE_PSRAM_CAP_V  0x00000003U
+# define EFUSE_PSRAM_CAP_S  3
 
 /** EFUSE_PSRAM_VENDOR : R; bitpos: [8:7]; default: 0;
  * register: EFUSE_RD_MAC_SPI_SYS_4_REG
  */
-  # define EFUSE_PSRAM_VENDOR    0x00000003U
-  # define EFUSE_PSRAM_VENDOR_M  (EFUSE_PSRAM_VENDOR_V << EFUSE_PSRAM_VENDOR_S)
-  # define EFUSE_PSRAM_VENDOR_V  0x00000003U
-  # define EFUSE_PSRAM_VENDOR_S  7
+# define EFUSE_PSRAM_VENDOR    0x00000003U
+# define EFUSE_PSRAM_VENDOR_M  (EFUSE_PSRAM_VENDOR_V << EFUSE_PSRAM_VENDOR_S)
+# define EFUSE_PSRAM_VENDOR_V  0x00000003U
+# define EFUSE_PSRAM_VENDOR_S  7
 
 
 /*
-    switch (psram_vendor)
-    {
-      case 1: features += F("(AP_3v3)"); break;
-      case 2: features += F("(AP_1v8)"); break;
-    }
+   switch (psram_vendor)
+   {
+   case 1: features += F("(AP_3v3)"); break;
+   case 2: features += F("(AP_1v8)"); break;
+   }
  */
 int32_t getEmbeddedPSRAMSize()
 {
@@ -109,5 +111,47 @@ bool isPSRAMInterfacePin(int gpio) {
 
 # endif // ifndef isPSRAMInterfacePin
 
+
+const __FlashStringHelper* getChipModel(uint32_t chip_model, uint32_t chip_revision, uint32_t pkg_version, bool single_core)
+{
+  if ((CHIP_ESP32S3 == chip_model) || // ESP32-S3
+      (4 == chip_model) ||            // ESP32-S3(beta2)
+      (6 == chip_model)) {            // ESP32-S3(beta3)
+    /*
+       ESP32-S3 Series
+       - 32-bit MCU & 2.4 GHz Wi-Fi & Bluetooth 5 (LE)
+       - Xtensa® 32-bit LX7 dual-core processor that operates at up to 240 MHz
+       - 512 KB of SRAM and 384 KB of ROM on the chip, and SPI, Dual SPI, Quad SPI, Octal SPI, QPI, and OPI interfaces that allow connection
+         to flash and external RAM
+       - Additional support for vector instructions in the MCU, which provides acceleration for neural network computing and signal
+         processing workloads
+       - Peripherals include 45 programmable GPIOs, SPI, I2S, I2C, PWM, RMT, ADC and UART, SD/MMC host and TWAITM
+       - Reliable security features ensured by RSA-based secure boot, AES-XTS-based flash encryption, the innovative digital signature and
+         the HMAC peripheral, “World Controller”
+     */
+
+    /*
+
+       efuse_reg.h:
+       EFUSE_RD_MAC_SPI_SYS_0_REG = block1_addr
+       EFUSE_RD_MAC_SPI_SYS_3_REG = block1_addr + (4 * num_word)) // (num_word = 3)
+
+     */
+
+
+# if (ESP_IDF_VERSION_MAJOR >= 5)
+    pkg_version = REG_GET_FIELD(EFUSE_RD_MAC_SPI_SYS_3_REG, EFUSE_PKG_VERSION);
+
+    switch (pkg_version) {
+      case 0:              return F("ESP32-S3");        // QFN56
+      case 1:              return F("ESP32-S3-PICO-1"); // LGA56
+    }
+# endif // if (ESP_IDF_VERSION_MAJOR >= 5)
+
+    return F("ESP32-S3"); // Max 240MHz, Dual core, QFN 7*7, ESP32-S3-WROOM-1, ESP32-S3-DevKitC-1
+  }
+
+  return F("Unknown");
+}
 
 #endif // ifdef ESP32S3
